refactor(popups): move popups into a folder
